+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@ Heat Pump Display Release Notes
 ## In development
 
 - Add statistics screen to display some values of the DTA logging file of the controller.
+- Add compressor frequency.
 
 ## 16.0.1 (2022-03-11)